Divide ActivityCarouselView
